Add files via upload
The sound files dah.wav and dit.wav can be played along with the flashing LED on a raspberry pi. 
Various options have been added when morse.py is run from a terminal command line, such as providing a quoted text string to be converted to morse code, repeating a message, and suppressing the dah and dit sound files from being played. Make sure to pick up dah.wav and dit.wav and place them in the same folder as mores.py.